backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.2.3

# AI/ML Libraries
langchain==0.1.0
langchain-openai==0.0.6
langchain-community==0.0.15
openai>=1.10.0
tiktoken==0.5.2

# Vector Database
pinecone-client==5.0.1

# Job Market Data
google-search-results==2.4.2

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# HTTP & API
httpx==0.25.2
aiofiles==23.2.1

# Audio Processing
speechrecognition==3.10.0
pydub==0.25.1
librosa==0.10.1

# Background Tasks
celery==5.3.4
kombu==5.3.4

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.40.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-slugify==8.0.1
pytz==2023.3
dateutils==0.6.12
